fix deleted submission controller files
